requirements.txt
aiohttp==3.9.1
Pillow==10.1.0
qrcode[pil]==7.4.2
requests==2.31.0
aiofiles==23.2.1
python-dateutil==2.8.2
pydantic==2.5.2
asyncio-mqtt==0.16.2
